minor contracts refactor & clippy
