+++ contracts/security-nft/src/types.rs
@@ -18,6 +18,7 @@ pub type FreezeParams = concordium_cis2_security::FreezeParams<TokenId, TokenAmo
 pub type FreezeParam = concordium_cis2_security::FreezeParam<TokenId, TokenAmount>;
 pub type FrozenParams = concordium_cis2_security::FrozenParams<TokenId>;
 pub type FrozenResponse = concordium_cis2_security::FrozenResponse<TokenAmount>;
+pub use cis2_types::ContractMetadataUrl;
 pub use concordium_cis2_security::RecoverParam;
 
 #[derive(Serialize, SchemaType)]
@@ -25,32 +26,6 @@ pub struct InitParam {
     pub identity_registry: ContractAddress,
     pub compliance:        ContractAddress,
     pub sponsors:          Vec<ContractAddress>,
-}
-
-/// Represents the metadata URL and hash of a token.
-#[derive(SchemaType, Serial, Clone, Deserial)]
-pub struct ContractMetadataUrl {
-    pub url:  String,
-    pub hash: Option<String>,
-}
-
-impl From<ContractMetadataUrl> for MetadataUrl {
-    fn from(val: ContractMetadataUrl) -> Self {
-        MetadataUrl {
-            url:  val.url,
-            hash: {
-                if let Some(hash) = val.hash {
-                    let mut hash_bytes = [0u8; 32];
-                    match hex::decode_to_slice(hash, &mut hash_bytes) {
-                        Ok(_) => Some(hash_bytes),
-                        Err(_) => None,
-                    }
-                } else {
-                    None
-                }
-            },
-        }
-    }
 }
 
 #[derive(Serialize, SchemaType)]
