+++ contracts/security-sft/src/types.rs
@@ -1,6 +1,5 @@
 use super::error::Error;
 use concordium_cis2::{OnReceivingCis2Params, Receiver, TokenIdVec};
-pub use concordium_rwa_utils::cis2_conversions::Rate;
 use concordium_rwa_utils::{cis2_schema_types, cis2_types, concordium_cis2_security};
 use concordium_std::*;
 
@@ -23,39 +22,15 @@ pub type FreezeParams = concordium_cis2_security::FreezeParams<TokenId, TokenAmo
 pub type FrozenParams = concordium_cis2_security::FrozenParams<TokenId>;
 pub type FrozenResponse = concordium_cis2_security::FrozenResponse<TokenAmount>;
 pub type DepositParams = OnReceivingCis2Params<NftTokenId, NftTokenAmount>;
+pub use cis2_types::ContractMetadataUrl;
 pub use concordium_cis2_security::RecoverParam;
+pub use concordium_rwa_utils::cis2_conversions::Rate;
 
 #[derive(Serialize, SchemaType)]
 pub struct InitParam {
     pub identity_registry: ContractAddress,
     pub compliance:        ContractAddress,
     pub sponsors:          Vec<ContractAddress>,
-}
-
-/// Represents the metadata URL and hash of a token.
-#[derive(SchemaType, Serial, Clone, Deserial)]
-pub struct ContractMetadataUrl {
-    pub url:  String,
-    pub hash: Option<String>,
-}
-
-impl From<ContractMetadataUrl> for MetadataUrl {
-    fn from(val: ContractMetadataUrl) -> Self {
-        MetadataUrl {
-            url:  val.url,
-            hash: {
-                if let Some(hash) = val.hash {
-                    let mut hash_bytes = [0u8; 32];
-                    match hex::decode_to_slice(hash, &mut hash_bytes) {
-                        Ok(_) => Some(hash_bytes),
-                        Err(_) => None,
-                    }
-                } else {
-                    None
-                }
-            },
-        }
-    }
 }
 
 #[derive(Serialize, SchemaType)]
